+++ src/myprocess.cpp
@@ -17,6 +17,7 @@
 */
 
 #include "myprocess.h"
+#include <cinttypes>
 
 #ifdef Q_OS_WIN
 
@@ -158,7 +159,7 @@ void MyProcess::procFinished() {
 	qDebug("MyProcess::procFinished");
 
 #if !USE_TEMP_FILE
-	qDebug("MyProcess::procFinished: Bytes available: %ld", bytesAvailable());
+	qDebug("MyProcess::procFinished: Bytes available: %"PRId64, (int64_t)bytesAvailable());
 	if ( bytesAvailable() > 0 ) readStdOut();
 #else
 	timer.stop();
